feat(updated): add useCheckEmail & useEmailValidation & useNewPassword hooks
