+++ src/screens/CheckEmail/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useCheckEmail} from './useCheckEmail';
