+++ src/screens/CheckEmail/hooks/useCheckEmail.tsx
@@ -1,0 +1,25 @@
+import {useState} from 'react';
+import {useRequestLoading} from 'core/hooks';
+import {useDispatch} from 'react-redux';
+import {useNavigation} from '@react-navigation/native';
+import {checkUserEmailRequest} from 'core/reducers';
+import {CheckEmailScreenNavigationProps} from '../types';
+
+export const useCheckEmail = () => {
+  const navigation = useNavigation<CheckEmailScreenNavigationProps>();
+  const dispatch = useDispatch();
+  const [email, setEmail] = useState('');
+  const [isEmailCorrect, setIsEmailCorrect] = useState(false);
+
+  const {loading} = useRequestLoading(checkUserEmailRequest);
+
+  return {
+    navigation,
+    email,
+    setIsEmailCorrect,
+    dispatch,
+    isEmailCorrect,
+    loading,
+    setEmail,
+  };
+};
