+++ src/screens/EmailValidation/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useEmailValidation} from './useEmailValidation'
